fix: variable name, add : advertisement helper
